bpf: encap: move & use DSR-GENEVE helpers
[ upstream commit d2b17f418ebe424fccd894cba079998e89568333 ]

Locate the set_geneve_dsr_opt*() closer to the actual struct definition.
This needs shifting around some of the IPv6 helpers as well.

Also make use of the DSR-GENEVE helpers for the BPF tests. Here we actually
need to initialize the `port` field, so that nodeport_dsr_ingress_ipv*()
doesn't drop the packet with DROP_INVALID. Curious that this as has worked
until now.
