+++ bpf/lib/ipv6.h
@@ -171,27 +171,6 @@ static __always_inline int ipv6_hdrlen(struct __ctx_buff *ctx, __u8 *nexthdr)
 	return ipv6_hdrlen_offset(ctx, ETH_HLEN, nexthdr, NULL);
 }
 
-static __always_inline void ipv6_addr_copy(union v6addr *dst,
-					   const union v6addr *src)
-{
-	memcpy(dst, src, sizeof(*dst));
-}
-
-static __always_inline void ipv6_addr_copy_unaligned(union v6addr *dst,
-						     const union v6addr *src)
-{
-	dst->d1 = src->d1;
-	dst->d2 = src->d2;
-}
-
-static __always_inline bool ipv6_addr_equals(const union v6addr *a,
-					     const union v6addr *b)
-{
-	if (a->d1 != b->d1)
-		return false;
-	return a->d2 == b->d2;
-}
-
 static __always_inline
 void ipv6_sol_mc_mac_set(const union v6addr *addr, union macaddr *mac)
 {
